requirements.txt
# astroid==1.3.6
autopep8>=1.4
# logilab-common==1.4.2
pep8>=1.7.1
# pylint==2.0.1
# six==1.11.0
# numpy==1.15.0
# biopython==1.72
PyYAML>=3.13
requests>=2.19
